split common enum files
