+++ html-exercise/typescripts/api/apiUrl.ts
@@ -1,0 +1,3 @@
+export enum endpoints {
+  PRODUCTS = '/html-exercise/productList.json'
+}
