--- html-exercise/typescripts/cart/cart.interface.ts
@@ -1,5 +1,0 @@
-import { ProductProps } from "../product/product.interface";
-
-export default interface CartItemProps extends Omit<ProductProps, 'status'> {
-  quantity: number
-}
